Initial Bug Fixes
Warning: rubbos.properties_1000 is the current config being used and it does not have a template meaning you need to hand replace "azhang" with your username. Also the constraints under docker-compose-swarm.yml.template might be different depending on the node you choose, specifically if it is infosphere or infosphere-pg0. The docker-compose-swarm.yml.template file in this commit was tested heavily with the emulab nodes.